+++ CHANGELOG.md
@@ -7,9 +7,11 @@ TODO
 Changelog
 =========
 
-##### Version 1.02.002 (WIP)
+##### Version 1.03.001 (WIP)
  * fix bugs
  * keep working
+ * add total projects dashboard widget
+ * add total contact messages dashboard widget
 
 ##### Version 1.02.001 (2023-0618)
  * add Enterprise management